news_graphene/requirements.txt
Package            Version
------------------ -------
aniso8601          7.0.0
asgiref            3.5.0
Django             3.2.13
django-graphql-jwt 0.3.4
graphene           2.1.9
graphene-django    2.15.0
graphql-core       2.3.2
graphql-relay      2.0.1
Pillow             9.1.0
pip                22.0.4
promise            2.3
PyJWT              2.3.0
pytz               2022.1
Rx                 1.6.1
setuptools         61.0.0
singledispatch     3.7.0
six                1.16.0
sqlparse           0.4.2
text-unidecode     1.3
wheel              0.37.1
